djvu2pdf_toc_parser: Use Python 3 as the interpreter.
The script is fully compatible with Python 3 (tested) and works well even
with some input files that had UTF-8 characters in them.

Characters in latin-1 make the script raise an exception, which means that
the script should probably treat its input as bytes.

That being said, some new operating systems do not have a Python 2
interpreter (and the situation will only get increasingly harder in the
future).